requirements.txt
streamlit
pandas
plotly
requests
supabase==2.0.0
python-dotenv